Move lobby challenge tests to test_lobby_challenge.py
